- switched to vectors for zonestatus on DSC - used static std::string to store persistent vales on binary_sensor

+++ components/vista_alarm_panel/vistaalarm.h
@@ -375,7 +375,7 @@ void setup() override;
     void alarm_keypress(std::string keystring);
 
     void alarm_keypress_partition(std::string keystring, int partition);
-    void send_cmd_bytes(uint8_t addr,std::string hexbytes);
+    void send_cmd_bytes(int addr,std::string hexbytes);
     void setDefaultKpAddr(uint8_t p);
     
 private:
